fix key size

+++ aqo_shared.c
@@ -242,7 +242,7 @@ aqo_init_shmem(void)
 								 &info, HASH_ELEM | HASH_BLOBS);
 
 	/* Shared memory hash table for fss neighbours */
-	info.keysize = sizeof(int);
+	info.keysize = sizeof(int64);
 	info.entrysize = sizeof(DataEntry *);
 	fss_neighbours = ShmemInitHash("AQO fss neighbours HTAB", fss_max_items, fss_max_items,
 							  &info, HASH_ELEM | HASH_BLOBS);
